Added Camera app
aswell as its icon and localisation aswell as updated some assistant key terms and fixed a long lasting persistant bug in the windowing system causing Desktops to error out and throw errors whilst trying to use the taskbar. Also added the option to replace your Desktop with a terminal by porting the terminal app to a ui app.